+++ balkan/sun.c
@@ -1,9 +1,11 @@
 /* Sun style partitioning */
 
+#include "balkan.h"
+
 #include <fcntl.h>
 #include <unistd.h>
+#include <sys/types.h>
 
-#include "balkan.h"
 #include "byteswap.h"
 
 struct singlePartitionTable {
@@ -98,7 +100,7 @@ int sunpReadTable(int fd, struct partitionTable * table) {
 	  default:
 	    if (table->parts[i].type != WHOLE_DISK &&
 		lseek64(fd, (8192 + 0x55c + SECTOR_SIZE *
-			    (unsigned long long)table->parts[i].startSector),
+			    (off64_t) table->parts[i].startSector),
 		       SEEK_SET) >= 0 &&
 		read(fd, &magic, 4) == 4 &&
 		(magic == UFS_SUPER_MAGIC ||
